requirements.txt
torch
torchvision
streamlit
numpy
pandas
pillow
matplotlib
